Move PrepareServiceInstance func to testutil package
