+++ test/service_instance_test/service_instance_test.go
@@ -19,6 +19,7 @@ package service_test
 import (
 	"context"
 	"fmt"
+	"github.com/Peripli/service-manager/test/testutil"
 
 	"github.com/Peripli/service-manager/pkg/query"
 	"github.com/Peripli/service-manager/pkg/types"
@@ -87,7 +88,7 @@ var _ = test.DescribeTestsFor(test.TestCase{
 
 				When("service instance contains tenant identifier in OSB context", func() {
 					BeforeEach(func() {
-						_, serviceInstance = test.PrepareServiceInstance(ctx, ctx.TestPlatform.ID, "", fmt.Sprintf(`{"%s":"%s"}`, TenantIdentifier, TenantValue))
+						_, serviceInstance = testutil.PrepareServiceInstance(ctx, ctx.TestPlatform.ID, "", fmt.Sprintf(`{"%s":"%s"}`, TenantIdentifier, TenantValue))
 						_, err := ctx.SMRepository.Create(context.Background(), serviceInstance)
 						Expect(err).ToNot(HaveOccurred())
 					})
@@ -101,7 +102,7 @@ var _ = test.DescribeTestsFor(test.TestCase{
 				})
 				When("service instance doesn't contain tenant identifier in OSB context", func() {
 					BeforeEach(func() {
-						_, serviceInstance = test.PrepareServiceInstance(ctx, ctx.TestPlatform.ID, "", "{}")
+						_, serviceInstance = testutil.PrepareServiceInstance(ctx, ctx.TestPlatform.ID, "", "{}")
 						_, err := ctx.SMRepository.Create(context.Background(), serviceInstance)
 						Expect(err).ToNot(HaveOccurred())
 					})
@@ -125,7 +126,7 @@ var _ = test.DescribeTestsFor(test.TestCase{
 })
 
 func blueprint(ctx *common.TestContext, auth *common.SMExpect) common.Object {
-	_, serviceInstance := test.PrepareServiceInstance(ctx, ctx.TestPlatform.ID, "", fmt.Sprintf(`{"%s":"%s"}`, TenantIdentifier, TenantValue))
+	_, serviceInstance := testutil.PrepareServiceInstance(ctx, ctx.TestPlatform.ID, "", fmt.Sprintf(`{"%s":"%s"}`, TenantIdentifier, TenantValue))
 	_, err := ctx.SMRepository.Create(context.Background(), serviceInstance)
 	if err != nil {
 		Fail(fmt.Sprintf("could not create service instance: %s", err))
